+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 ### Added
 - A new option is added to pad zero for timestamps copied.
+- A reset button is added to the option UI.
 
 ## [1.2.0] - 2025-02-18
 ### Added